refactor: cleanup main.rs
