+++ cli/tools/mod.rs
@@ -1,6 +1,7 @@
 // Copyright 2018-2022 the Deno authors. All rights reserved. MIT license.
 
 pub mod bench;
+pub mod bundle;
 pub mod check;
 pub mod coverage;
 pub mod doc;
@@ -10,6 +11,7 @@ pub mod init;
 pub mod installer;
 pub mod lint;
 pub mod repl;
+pub mod run;
 pub mod standalone;
 pub mod task;
 pub mod test;
